divide the e2e test into multiple files
