+++ tests/e2e/functional_test.go
@@ -18,33 +18,26 @@ package e2e
 
 import (
 	"encoding/json"
-	"fmt"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/client-go/util/retry"
 
 	numaflowv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
 
-	"github.com/numaproj/numaplane/internal/util"
-	"github.com/numaproj/numaplane/internal/util/kubernetes"
 	apiv1 "github.com/numaproj/numaplane/pkg/apis/numaplane/v1alpha1"
 )
 
 const (
-	numaflowControllerRolloutName = "numaflow-controller"
-	isbServiceRolloutName         = "test-isbservice-rollout"
-	isbServiceStatefulSetName     = "isbsvc-test-isbservice-rollout-js"
-	pipelineRolloutName           = "test-pipeline-rollout"
+	isbServiceRolloutName     = "test-isbservice-rollout"
+	isbServiceStatefulSetName = "isbsvc-test-isbservice-rollout-js"
+	pipelineRolloutName       = "test-pipeline-rollout"
 )
 
 var (
@@ -98,7 +91,7 @@ func init() {
 	isbservicegvr = getGVRForISBService()
 }
 
-var _ = Describe("PipelineRollout e2e", func() {
+var _ = Describe("Functional e2e", Serial, func() {
 
 	It("Should create the NumaflowControllerRollout if it doesn't exist", func() {
 
@@ -158,7 +151,7 @@ var _ = Describe("PipelineRollout e2e", func() {
 		document("Updating Pipeline directly")
 
 		// update child Pipeline
-		updatePipelineSpecInK8S(Namespace, pipelineRolloutName, func(pipelineSpec numaflowv1.PipelineSpec) (numaflowv1.PipelineSpec, error) {
+		updatePipelineSpecInK8S(pipelineRolloutName, func(pipelineSpec numaflowv1.PipelineSpec) (numaflowv1.PipelineSpec, error) {
 			rpu := int64(10)
 			pipelineSpec.Vertices[0].Source.Generator.RPU = &rpu
 			return pipelineSpec, nil
@@ -190,7 +183,7 @@ var _ = Describe("PipelineRollout e2e", func() {
 		Expect(err).ShouldNot(HaveOccurred())
 
 		// update the PipelineRollout
-		updatePipelineRolloutInK8S(Namespace, pipelineRolloutName, func(rollout apiv1.PipelineRollout) (apiv1.PipelineRollout, error) {
+		updatePipelineRolloutInK8S(pipelineRolloutName, func(rollout apiv1.PipelineRollout) (apiv1.PipelineRollout, error) {
 			rollout.Spec.Pipeline.Spec.Raw = rawSpec
 			return rollout, nil
 		})
@@ -360,179 +353,6 @@ var _ = Describe("PipelineRollout e2e", func() {
 
 })
 
-// document for Ginkgo framework and print to console
-func document(testName string) {
-	snapshotCluster(testName)
-	By(testName)
-}
-
-func snapshotCluster(testName string) {
-	fmt.Printf("*** %+v: NAMESPACE POD STATE BEFORE TEST: %s\n", time.Now(), testName)
-	podList, _ := kubeClient.CoreV1().Pods(Namespace).List(ctx, metav1.ListOptions{})
-	if podList != nil {
-		for _, pod := range podList.Items {
-			fmt.Printf("Pod: %q, %q, Reason:%q\n", pod.Name, pod.Status.Phase, pod.Status.Reason)
-
-		}
-	}
-}
-
-// verify that the Deployment matches some criteria
-func verifyNumaflowControllerDeployment(namespace string, f func(appsv1.Deployment) bool) {
-	document("verifying Numaflow Controller Deployment")
-	Eventually(func() bool {
-		deployment, err := kubeClient.AppsV1().Deployments(namespace).Get(ctx, numaflowControllerRolloutName, metav1.GetOptions{})
-		if err != nil {
-			return false
-		}
-		return f(*deployment)
-	}).WithTimeout(testTimeout).Should(BeTrue())
-}
-
-func verifyNumaflowControllerReady(namespace string) {
-	document("Verifying that the Numaflow Controller Deployment exists")
-	Eventually(func() error {
-		_, err := kubeClient.AppsV1().Deployments(namespace).Get(ctx, numaflowControllerRolloutName, metav1.GetOptions{})
-		return err
-	}).WithTimeout(testTimeout).Should(Succeed())
-
-	document("Verifying that the Numaflow ControllerRollout is ready")
-	Eventually(func() bool {
-		rollout, _ := numaflowControllerRolloutClient.Get(ctx, numaflowControllerRolloutName, metav1.GetOptions{})
-		if rollout == nil {
-			return false
-		}
-		childResourcesHealthyCondition := rollout.Status.GetCondition(apiv1.ConditionChildResourceHealthy)
-		if childResourcesHealthyCondition == nil {
-			return false
-		}
-		return childResourcesHealthyCondition.Status == metav1.ConditionTrue
-
-	}).WithTimeout(testTimeout).Should(BeTrue())
-}
-
-func verifyISBServiceSpec(namespace string, name string, f func(numaflowv1.InterStepBufferServiceSpec) bool) {
-
-	document("verifying ISBService Spec")
-	var retrievedISBServiceSpec numaflowv1.InterStepBufferServiceSpec
-	Eventually(func() bool {
-		unstruct, err := dynamicClient.Resource(getGVRForISBService()).Namespace(namespace).Get(ctx, name, metav1.GetOptions{})
-		if err != nil {
-			return false
-		}
-		if retrievedISBServiceSpec, err = getISBServiceSpec(unstruct); err != nil {
-			return false
-		}
-
-		return f(retrievedISBServiceSpec)
-	}).WithTimeout(testTimeout).Should(BeTrue())
-}
-
-func verifyISBSvcReady(namespace string, isbsvc string, nodeSize int) {
-	document("Verifying that the ISBService exists")
-	Eventually(func() error {
-		_, err := dynamicClient.Resource(getGVRForISBService()).Namespace(namespace).Get(ctx, isbServiceRolloutName, metav1.GetOptions{})
-		return err
-	}).WithTimeout(testTimeout).Should(Succeed())
-
-	// TODO: eventually we can use ISBServiceRollout.Status.Conditions(ChildResourcesHealthy) to get this instead
-	document("Verifying that the StatefulSet exists and is ready")
-	Eventually(func() bool {
-		statefulSet, _ := kubeClient.AppsV1().StatefulSets(namespace).Get(ctx, isbServiceStatefulSetName, metav1.GetOptions{})
-		return statefulSet != nil && statefulSet.Generation == statefulSet.Status.ObservedGeneration && statefulSet.Status.UpdatedReplicas == int32(nodeSize)
-	}).WithTimeout(testTimeout).Should(BeTrue())
-
-	document("Verifying that the StatefulSet Pods are in Running phase")
-	Eventually(func() bool {
-		podList, _ := kubeClient.CoreV1().Pods(namespace).List(ctx, metav1.ListOptions{LabelSelector: fmt.Sprintf("%s=%s", numaflowv1.KeyISBSvcName, isbsvc)})
-		podsInRunning := 0
-		if podList != nil {
-			for _, pod := range podList.Items {
-				if pod.Status.Phase == corev1.PodRunning {
-					podsInRunning += 1
-				}
-			}
-		}
-		return podsInRunning == nodeSize
-	}).WithTimeout(testTimeout).Should(BeTrue())
-}
-
-func verifyPipelineSpec(namespace string, pipelineName string, f func(numaflowv1.PipelineSpec) bool) {
-
-	document("verifying Pipeline Spec")
-	var retrievedPipelineSpec numaflowv1.PipelineSpec
-	Eventually(func() bool {
-		unstruct, err := dynamicClient.Resource(getGVRForPipeline()).Namespace(namespace).Get(ctx, pipelineName, metav1.GetOptions{})
-		if err != nil {
-			return false
-		}
-		if retrievedPipelineSpec, err = getPipelineSpec(unstruct); err != nil {
-			return false
-		}
-
-		return f(retrievedPipelineSpec)
-	}).WithTimeout(testTimeout).Should(BeTrue())
-}
-
-func verifyPipelineStatus(namespace string, pipelineName string, f func(numaflowv1.PipelineSpec, kubernetes.GenericStatus) bool) {
-
-	document("verifying PipelineStatus")
-	var retrievedPipelineSpec numaflowv1.PipelineSpec
-	var retrievedPipelineStatus kubernetes.GenericStatus
-	Eventually(func() bool {
-		unstruct, err := dynamicClient.Resource(getGVRForPipeline()).Namespace(namespace).Get(ctx, pipelineName, metav1.GetOptions{})
-		if err != nil {
-			return false
-		}
-		if retrievedPipelineSpec, err = getPipelineSpec(unstruct); err != nil {
-			return false
-		}
-		if retrievedPipelineStatus, err = getNumaflowResourceStatus(unstruct); err != nil {
-			return false
-		}
-
-		return f(retrievedPipelineSpec, retrievedPipelineStatus)
-	}).WithTimeout(testTimeout).Should(BeTrue())
-}
-
-func verifyPipelineReady(namespace string, pipelineName string, numVertices int) {
-	document("Verifying that the Pipeline is running")
-	verifyPipelineStatus(namespace, pipelineName,
-		func(retrievedPipelineSpec numaflowv1.PipelineSpec, retrievedPipelineStatus kubernetes.GenericStatus) bool {
-			return retrievedPipelineStatus.Phase == string(numaflowv1.PipelinePhaseRunning)
-		})
-
-	vertexLabelSelector := fmt.Sprintf("%s=%s,%s=%s", numaflowv1.KeyPipelineName, pipelineName, numaflowv1.KeyComponent, "vertex")
-	daemonLabelSelector := fmt.Sprintf("%s=%s,%s=%s", numaflowv1.KeyPipelineName, pipelineName, numaflowv1.KeyComponent, "daemon")
-
-	// Get Pipeline Pods to verify they're all up
-	// TODO: eventually we can use PipelineRollout.Status.Conditions(ChildResourcesHealthy) to get this instead
-	document("Verifying that the Pipeline is ready")
-	// check "vertex" Pods
-	verifyPodsRunning(namespace, 2, vertexLabelSelector)
-	verifyPodsRunning(namespace, 1, daemonLabelSelector)
-
-}
-
-func verifyPodsRunning(namespace string, numPods int, labelSelector string) {
-	document(fmt.Sprintf("verifying %d Pods running with label selector %q", numPods, labelSelector))
-
-	Eventually(func() bool {
-		podsList, _ := kubeClient.CoreV1().Pods(namespace).List(ctx, metav1.ListOptions{LabelSelector: labelSelector})
-		if podsList != nil && len(podsList.Items) >= numPods {
-			for _, pod := range podsList.Items {
-				if pod.Status.Phase != "Running" {
-					return false
-				}
-			}
-			return true
-		}
-		return false
-
-	}).WithTimeout(testTimeout).Should(BeTrue())
-
-}
-
 func createPipelineRolloutSpec(name, namespace string) *apiv1.PipelineRollout {
 
 	pipelineSpecRaw, err := json.Marshal(pipelineSpec)
@@ -558,123 +378,6 @@ func createPipelineRolloutSpec(name, namespace string) *apiv1.PipelineRollout {
 
 	return pipelineRollout
 
-}
-
-func updateNumaflowControllerRolloutInK8S(f func(apiv1.NumaflowControllerRollout) (apiv1.NumaflowControllerRollout, error)) {
-	document("updating NumaflowControllerRollout")
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		rollout, err := numaflowControllerRolloutClient.Get(ctx, numaflowControllerRolloutName, metav1.GetOptions{})
-		if err != nil {
-			return err
-		}
-
-		*rollout, err = f(*rollout)
-		if err != nil {
-			return err
-		}
-		_, err = numaflowControllerRolloutClient.Update(ctx, rollout, metav1.UpdateOptions{})
-		return err
-	})
-	Expect(err).ShouldNot(HaveOccurred())
-}
-
-func updateISBServiceRolloutInK8S(name string, f func(apiv1.ISBServiceRollout) (apiv1.ISBServiceRollout, error)) {
-	document("updating ISBServiceRollout")
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		rollout, err := isbServiceRolloutClient.Get(ctx, name, metav1.GetOptions{})
-		if err != nil {
-			return err
-		}
-
-		*rollout, err = f(*rollout)
-		if err != nil {
-			return err
-		}
-		_, err = isbServiceRolloutClient.Update(ctx, rollout, metav1.UpdateOptions{})
-		return err
-	})
-	Expect(err).ShouldNot(HaveOccurred())
-}
-
-func updatePipelineRolloutInK8S(namespace string, name string, f func(apiv1.PipelineRollout) (apiv1.PipelineRollout, error)) {
-	document("updating PipelineRollout")
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		rollout, err := pipelineRolloutClient.Get(ctx, pipelineRolloutName, metav1.GetOptions{})
-		if err != nil {
-			return err
-		}
-
-		*rollout, err = f(*rollout)
-		if err != nil {
-			return err
-		}
-		_, err = pipelineRolloutClient.Update(ctx, rollout, metav1.UpdateOptions{})
-		return err
-	})
-	Expect(err).ShouldNot(HaveOccurred())
-}
-
-func updatePipelineSpecInK8S(namespace string, pipelineName string, f func(numaflowv1.PipelineSpec) (numaflowv1.PipelineSpec, error)) {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-
-		unstruct, err := dynamicClient.Resource(pipelinegvr).Namespace(Namespace).Get(ctx, pipelineRolloutName, metav1.GetOptions{})
-		Expect(err).ShouldNot(HaveOccurred())
-		retrievedPipeline := unstruct
-
-		// modify Pipeline Spec to verify it gets auto-healed
-		err = updatePipelineSpec(retrievedPipeline, f)
-		Expect(err).ShouldNot(HaveOccurred())
-		_, err = dynamicClient.Resource(pipelinegvr).Namespace(Namespace).Update(ctx, retrievedPipeline, metav1.UpdateOptions{})
-		return err
-	})
-	Expect(err).ShouldNot(HaveOccurred())
-}
-
-// Take a Pipeline Unstructured type and update the PipelineSpec in some way
-func updatePipelineSpec(u *unstructured.Unstructured, f func(numaflowv1.PipelineSpec) (numaflowv1.PipelineSpec, error)) error {
-
-	// get PipelineSpec from unstructured object
-	specMap := u.Object["spec"]
-	var pipelineSpec numaflowv1.PipelineSpec
-	err := util.StructToStruct(&specMap, &pipelineSpec)
-	if err != nil {
-		return err
-	}
-	// update PipelineSpec
-	pipelineSpec, err = f(pipelineSpec)
-	if err != nil {
-		return err
-	}
-	var newMap map[string]interface{}
-	err = util.StructToStruct(&pipelineSpec, &newMap)
-	if err != nil {
-		return err
-	}
-	u.Object["spec"] = newMap
-	return nil
-}
-
-// Get ISBServiceSpec from Unstructured type
-func getISBServiceSpec(u *unstructured.Unstructured) (numaflowv1.InterStepBufferServiceSpec, error) {
-	specMap := u.Object["spec"]
-	var isbServiceSpec numaflowv1.InterStepBufferServiceSpec
-	err := util.StructToStruct(&specMap, &isbServiceSpec)
-	return isbServiceSpec, err
-}
-
-// Get PipelineSpec from Unstructured type
-func getPipelineSpec(u *unstructured.Unstructured) (numaflowv1.PipelineSpec, error) {
-	specMap := u.Object["spec"]
-	var pipelineSpec numaflowv1.PipelineSpec
-	err := util.StructToStruct(&specMap, &pipelineSpec)
-	return pipelineSpec, err
-}
-
-func getNumaflowResourceStatus(u *unstructured.Unstructured) (kubernetes.GenericStatus, error) {
-	statusMap := u.Object["status"]
-	var status kubernetes.GenericStatus
-	err := util.StructToStruct(&statusMap, &status)
-	return status, err
 }
 
 func createNumaflowControllerRolloutSpec(name, namespace string) *apiv1.NumaflowControllerRollout {
@@ -722,20 +425,4 @@ func createISBServiceRolloutSpec(name, namespace string) *apiv1.ISBServiceRollou
 
 	return isbServiceRollout
 
-}
-
-func getGVRForPipeline() schema.GroupVersionResource {
-	return schema.GroupVersionResource{
-		Group:    "numaflow.numaproj.io",
-		Version:  "v1alpha1",
-		Resource: "pipelines",
-	}
-}
-
-func getGVRForISBService() schema.GroupVersionResource {
-	return schema.GroupVersionResource{
-		Group:    "numaflow.numaproj.io",
-		Version:  "v1alpha1",
-		Resource: "interstepbufferservices",
-	}
 }
