+++ tests/e2e/isbservice.go
@@ -1,0 +1,98 @@
+package e2e
+
+import (
+	"fmt"
+
+	. "github.com/onsi/gomega"
+	corev1 "k8s.io/api/core/v1"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/client-go/util/retry"
+
+	numaflowv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
+
+	"github.com/numaproj/numaplane/internal/util"
+	apiv1 "github.com/numaproj/numaplane/pkg/apis/numaplane/v1alpha1"
+)
+
+func getGVRForISBService() schema.GroupVersionResource {
+	return schema.GroupVersionResource{
+		Group:    "numaflow.numaproj.io",
+		Version:  "v1alpha1",
+		Resource: "interstepbufferservices",
+	}
+}
+
+// Get ISBServiceSpec from Unstructured type
+func getISBServiceSpec(u *unstructured.Unstructured) (numaflowv1.InterStepBufferServiceSpec, error) {
+	specMap := u.Object["spec"]
+	var isbServiceSpec numaflowv1.InterStepBufferServiceSpec
+	err := util.StructToStruct(&specMap, &isbServiceSpec)
+	return isbServiceSpec, err
+}
+
+func verifyISBServiceSpec(namespace string, name string, f func(numaflowv1.InterStepBufferServiceSpec) bool) {
+
+	document("verifying ISBService Spec")
+	var retrievedISBServiceSpec numaflowv1.InterStepBufferServiceSpec
+	Eventually(func() bool {
+		unstruct, err := dynamicClient.Resource(getGVRForISBService()).Namespace(namespace).Get(ctx, name, metav1.GetOptions{})
+		if err != nil {
+			return false
+		}
+		if retrievedISBServiceSpec, err = getISBServiceSpec(unstruct); err != nil {
+			return false
+		}
+
+		return f(retrievedISBServiceSpec)
+	}).WithTimeout(testTimeout).Should(BeTrue())
+}
+
+func verifyISBSvcReady(namespace string, isbsvcName string, nodeSize int) {
+	document("Verifying that the ISBService exists")
+	Eventually(func() error {
+		_, err := dynamicClient.Resource(getGVRForISBService()).Namespace(namespace).Get(ctx, isbsvcName, metav1.GetOptions{})
+		return err
+	}).WithTimeout(testTimeout).Should(Succeed())
+
+	// TODO: eventually we can use ISBServiceRollout.Status.Conditions(ChildResourcesHealthy) to get this instead
+	document("Verifying that the StatefulSet exists and is ready")
+	Eventually(func() bool {
+		statefulSet, _ := kubeClient.AppsV1().StatefulSets(namespace).Get(ctx, fmt.Sprintf("isbsvc-%s-js", isbsvcName), metav1.GetOptions{})
+		return statefulSet != nil && statefulSet.Generation == statefulSet.Status.ObservedGeneration && statefulSet.Status.UpdatedReplicas == int32(nodeSize)
+	}).WithTimeout(testTimeout).Should(BeTrue())
+
+	document("Verifying that the StatefulSet Pods are in Running phase")
+	Eventually(func() bool {
+		podList, _ := kubeClient.CoreV1().Pods(namespace).List(ctx, metav1.ListOptions{LabelSelector: fmt.Sprintf("%s=%s", numaflowv1.KeyISBSvcName, isbsvcName)})
+		podsInRunning := 0
+		if podList != nil {
+			for _, pod := range podList.Items {
+				if pod.Status.Phase == corev1.PodRunning {
+					podsInRunning += 1
+				}
+			}
+		}
+		return podsInRunning == nodeSize
+	}).WithTimeout(testTimeout).Should(BeTrue())
+}
+
+func updateISBServiceRolloutInK8S(name string, f func(apiv1.ISBServiceRollout) (apiv1.ISBServiceRollout, error)) {
+	document("updating ISBServiceRollout")
+	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		rollout, err := isbServiceRolloutClient.Get(ctx, name, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
+
+		*rollout, err = f(*rollout)
+		if err != nil {
+			return err
+		}
+		_, err = isbServiceRolloutClient.Update(ctx, rollout, metav1.UpdateOptions{})
+		return err
+	})
+	Expect(err).ShouldNot(HaveOccurred())
+}
